Move mem tile specific code into appropriate class
